mmio: fix duplicate transactions in integral read/write methods
the integral write methods (e.g. write32, write16, write8) were causing
duplicate writes on the memory bus, apparently due to the underlying
implementation of packing / unpacking a memory mapping slice. this
commit modifies the integral read and write methods to dereference a
ctypes pointer of the appropriate width instead, which was tested and
shown to only yield one transaction on the bus. special thanks to
@paul-demo for raising this issue, investigating solutions, and testing.

resolves #50.